#81: UITextData Dialog 개선
