+++ src/types/components.ts
@@ -1,4 +1,10 @@
-import type { ReactElement, PropsWithChildren, SyntheticEvent } from 'react';
+import type {
+	ReactElement,
+	PropsWithChildren,
+	SyntheticEvent,
+	Dispatch,
+	SetStateAction,
+} from 'react';
 import type { NextPage } from 'next';
 import type { SxProps } from '@mui/system';
 import type { TUITag, IUITextData } from 'src/types/ui-data';
@@ -10,6 +16,10 @@ export type TNextPageWithLayout = NextPage & {
 
 export type TNavItem = {
 	items: { name: string; link: string; icon?: ReactElement }[];
+};
+
+export type TDataTableRow = {
+	[key: string]: string | number | boolean | ReactElement;
 };
 
 /* Interface */
@@ -66,11 +76,13 @@ export interface ISelectFilter {
 }
 
 export interface IUIDialogViewer {
+	write?: boolean;
 	open: boolean;
 	onClose: () => void;
-	ImageComponent?: ReactElement;
-	HeaderComponent?: ReactElement;
-	TextComponent?: ReactElement | string;
+	tags?: IUITagComponents;
+	data?: IUITextData;
+	setData?: Dispatch<SetStateAction<IUITextData>>;
+	onUITextCopy?: (id: number, tooltip: (message: string) => void) => Promise<void>;
 	BottomComponent?: ReactElement;
 }
